# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: BookingHub API
  description: Multi-tenant SaaS booking platform API
  version: 1.0.0
servers:
  - url: /api/v1
    description: Current API version
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /bookings:
    post:
      summary: Create a booking
      tags: [Bookings]
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Booking created
    get:
      summary: List user/company bookings
      tags: [Bookings]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of bookings
  /bookings/{id}:
    get:
      summary: Get booking details
      tags: [Bookings]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking details
    patch:
      summary: Update a booking (reschedule)
      tags: [Bookings]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking updated
    delete:
      summary: Cancel a booking
      tags: [Bookings]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Booking cancelled
...
  /analytics/bookings:
    get:
      summary: Get booking analytics
      tags: [Analytics]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Booking statistics
...